Port L1+L2 to 6502 asm: screen, disk, expr
Replace three C modules with hand-written 6502 assembly:
- screen.c → screen.s (scroll, newline, cursor, color)
- disk.c → disk.s (KERNAL direct: SETLFS/SETNAM/OPEN/CLOSE/
  CHKIN/CHKOUT/CHRIN/CHROUT/LOAD/SAVE/READST/CLRCHN)
- expr.c → expr.s (hex literal parser)

Eliminates all 19 cc65 CBM wrapper modules (cbm_*.o, c_*.o).
Directory listing reads raw $-channel format directly.
PRG load/save consolidated into disk.s (removed from repl.c).

Binary: 22,627 → 20,628 bytes (−1,999 bytes, −8.8%).
All 2277 tests pass.

+++ src/disk.h
@@ -10,9 +10,11 @@ void floppy_status(void);
 /* Directory listing to screen. RUN/STOP breaks. */
 void list_directory(uint8_t device);
 
-/* PRG file I/O */
-uint8_t disk_load_prg(const char *name, uint16_t addr);
-uint8_t disk_save_prg(const char *name, uint16_t addr, uint16_t size);
+/* PRG file I/O.
+ * disk_load_prg returns end address (nonzero) on success, 0 on error.
+ * disk_save_prg returns 0 on success, nonzero on error. */
+uint16_t disk_load_prg(const char *name, uint16_t addr);
+uint8_t  disk_save_prg(const char *name, uint16_t addr, uint16_t size);
 
 /* SEQ file I/O — callback-based to avoid editor dependency.
  * insert_fn: called with each byte read from file.
